feat: Add ServiceJourneyWithEstimatedCalls
The main new feature is adding notices, however since this was already a
non-backward compatible change the old ServiceJourneyDepartures endpoint
was replaced with a new ServiceJourneyWithEstimatedCalls endpoint to fix
some illogical structuring of the old query.

Some examples:
- The old endppint retrieved serviceJourney fields through each
  estimated call. This is not necessary as the root element is the
  serviceJourney that every call references.
- The old endpoint accessed the line (for publicCode) through
  serviceJourney -> estimatedCall -> serviceJourney -> journeyPattern ->
  line, when it could be serviceJourney -> line.
- TransportMode and TransportSubmode were queried on line when it can be
  queried directly on the serviceJourney.

Also created reusable fragments for the fragments used in the new
endpoint.