fix: Polyline endpoint fail without cache
The polyline server method has a third parameter
'headers: Request<ReqRefDefaults>', but this third parameter wasn't
provided when invoking the server method from the api endpoint. So
the third parameter was actually resolved to the cache config
which hapi behind the scenes provide as an extra parameter to the
server method invocation. When no cache was set, this parameter was
undefined, which would give an error later in the request handling
in the bff.