feat: Disable the Apollo Client cache
The theory is that the Apollo Client cache is the reason why we are
seeing a large memory usage in prod. Also the client doesn't make as
much sense when we have multiple BFF instances running in prod.

The main usage for the Apollo Client cache was to cache the previous
result for realtime updates, so only calls that were actually
updated could be returned. This is now changed to only return
updates for calls up to 30 minutes in the future, and no usage of
cache.