feat: Add line id on service journey query
The line id is necessary when storing favorites on the departure
details screen. We also add line mode and submode to be stored as
metadata on the favorite. We already have mode and submode on the
service journey, but technically it is more correct to store line
modes since service journey modes may be overwritten by temporary
changes like bus for train.