feat: Remove stop id and make quay id required on favorites
Whoa! Isn't this a non-backwards compatible change and requires a
new v3 endpoint? The reason we modify the v2 endpoint is that by
looking at code and logs it seems that we always store and send in
quay id on favorites. And when we always have quay id we no longer
need stop place id, as the quay id is uniquely identifiable.

This change should be backwards compatible and return responsen in
the same format as before, but needs to be thoroughly tested before
production and observed after production.